fix(studio): extract pure helpers to studio_helpers.py to avoid Streamlit import at test time
Moves _compile_safe, _load_identity_safe, and _studio_agent_from_yaml out of
web/studio.py (which imports streamlit at module level) into a new
web/studio_helpers.py with no Streamlit dependency. Tests now import from
studio_helpers directly, unblocking collection in envs without Streamlit
installed. Streamlit remains an optional [web] extra as designed.

Fixes post-merge CI failure on test (3.13) — initiative 8dd9c199.
